refactor: update server entry point and adjust nodemon configuration
