+++ nodemon.json
@@ -2,5 +2,5 @@
   "watch": ["src"],
   "ext": ".ts,.js",
   "ignore": [],
-  "exec": "node --loader ts-node/esm src/index.ts"
+  "exec": "node --loader ts-node/esm src/server/index.ts"
 }
