+++ package.json
@@ -3,11 +3,11 @@
   "version": "1.0.0",
   "type": "module",
   "scripts": {
-    "start": "node dist/index.js",
+    "start": "node dist/server/index.js",
     "dev": "nodemon",
     "build": "tsc",
     "format": "prettier --write .",
-    "serve": "node dist/index.js",
+    "serve": "node dist/server/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
